# public/swagger.yaml
openapi: 3.0.4
info:
  title: API for WMS - Documented by Swagger
  description: |-
    This is an API documentation of all endpoints available for the use of WMS

  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
paths:
...
  /api/outbound-shipment:
    get:
      tags:
        - Outbound functions
      summary: Get all overall information of all outbound shipments
      description: |
        Retrieves all outbound shipments with their relevant overall features.
      responses:
        "200":
          description: A list of outbound shipments with related customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OutboundShipment"
    post:
      tags:
        - Outbound functions
      summary: Create a new outbound shipment
      description: Creates a new outbound shipment. All fields are optional and will use defaults if not provided.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shipmentDate:
                  type: string
                  format: date-time
                  nullable: true
                  example: "2025-04-24T10:00:00Z"
                  description: Optional shipment date. Defaults to current time.
                carrier:
                  type: string
                  maxLength: 100
                  nullable: true
                  example: "UPS"
                  description: Optional carrier name
                trackingNumber:
                  type: string
                  maxLength: 100
                  nullable: true
                  example: "1Z1234567890"
                  description: Optional tracking number. Must be unique if provided.
      responses:
        "200":
          description: Shipment successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Outbound shipment created successfully.
                  data:
                    type: object
                    example:
                      insertId: 12
        "500":
          description: Internal server error while inserting data
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error inserting data
    put:
      tags:
        - Outbound functions
    delete:
      tags:
        - Outbound functions
...
components:
  schemas:
...
    OutboundShipment:
      type: object
      required:
        - shipmentId
      properties:
        shipmentId:
          type: integer
          format: int32
          description: Auto-incremented primary key for the shipment
          example: 1
        shipmentDate:
          type: string
          format: date-time
          description: Date and time the shipment was created
          example: "2025-04-24T15:30:00Z"
        carrier:
          type: string
          maxLength: 100
          nullable: true
          description: Name of the carrier handling the shipment
          example: "DHL Express"
        trackingNumber:
          type: string
          maxLength: 100
          nullable: true
          description: Unique tracking number for the shipment
          example: "1Z9999999999999999"
...